test: move app.config defaultVariants regression into useComponentProps spec
